Clean up old prototype code
