+++ src/App.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
-// import Pigeonburgh from './Pigeonburgh';
-import Pigeonburgh2 from './Pigeonburgh2';
+import Pigeonburgh from './Pigeonburgh';
 
 class App extends Component {
 	render() {
 		return (
 			<div>
-				<Pigeonburgh2 />
+				<Pigeonburgh />
 			</div>
 		);
 	}
