--- src/substepText.js
@@ -1,27 +1,0 @@
-import React from 'react';
-
-export const stepOneSubstepOne = 'Yes, a pigeon. Stunning royal blue and gray feathers. Orange feetsies.';
-
-export const stepThreeSubstepOne =
-	"SIIIIIIIP. The lukewarm bev flows through you. It's the perfect ratio of cream to coffee. Now you are ready for action.";
-
-export const stepFourSubstepOne =
-	'Cooler than you. Wearing t-shirts featuring bands you’ve never heard of: Death by Plastic Monstertruck, Puddleslap, Figboy and the Newtons.';
-
-export const stepSixSubstepOne = 'Real classics like "I\'ll be back" and "Chill out, dickwad."';
-
-export const stepSixSubstepTwo = <img src="assets/maverick.gif" alt="Maverick" />;
-
-export const stepTwelveSubstepOne =
-	"Ira was cool for sure. She cracked the best jokes in still life illustration class. During nude drawing sessions, she called it the Art S'nudio. Ha. Classic.";
-
-export const stepThirteenSubstepOne = "A deli paradise. That's the kind of utopia I want to live in, forever.";
-
-export const stepFourteenSubstepOne =
-	'A metal butterfly destroying it’s Slayer chrysalis while double-bass drums resound in the background.';
-
-export const stepFifteenSubstepOne =
-	'Flashback to KAREN! DANG IT she doesn’t even know how to use Microsoft Paint don’t MESS WITH ME KAREN I SWEAR.';
-
-export const stepNineteenSubstepOne =
-	"Fluent in binary code, really though? Maybe you'll ask her to hang sometime to see what other stuff is going on in her steel trap of a brain. But she’s probably too cool and connected and maybe you’ll ask later and anyways you’re super busy solving this weird puzzle and and and and and AND";
